xbfwup: improve feedback as well as reliability
The time it takes to enter the bootloader seems highly variable,
ranging from as little as a second after the firmware reset to
almost four seconds.  Since we only waited for two seconds, we
would occasionally abort before the bootloader fully loaded.
Also, the xbee would pause for ~2 seconds after we sent the last
hunk of the new firmware, giving the appearance of a hang mid-flash.

Wait 10 seconds for the bootloader to come up, printing an indicator
while waiting.  Also print a message before reading the final ACK
from the xbee so that the pause isn't so alarming.